// src/types/experience.ts
// types/experience.ts
export interface IExperience {
  logo?: string
  isActive: boolean
  company: string
  role: string
  job: string
  start: string
  end: string
  location: string
  city: string
  country: string
  website?: string
  description: string
  responsibilities: string[]
  projectImage?: string
}